Refactor config and allow dynamic config creation
