+++ zen_creator/__init__.py
@@ -21,7 +21,7 @@ from .model import Model
 from .sectors import Sector
 from .utils.attribute import Attribute
 from .utils.compare_trees import compare_trees
-from .utils.default_config import Config
+from .utils.config import Config, DatasetConfig, DatasetCollectionConfig, TechnologyConfig, CarrierConfig, ConversionTechnologyConfig, TransportTechnologyConfig, StorageTechnologyConfig
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
@@ -39,9 +39,16 @@ __all__ = [
     "StorageTechnology",
     "TransportTechnology",
     "Dataset",
+    "DatasetConfig",
     "MetaData",
     "DatasetCollection",
+    "DatasetCollectionConfig",
     "TechnoEconomicDataset",
     "Attribute",
     "SourceInformation",
+    "TechnologyConfig",
+    "CarrierConfig",
+    "TransportTechnologyConfig",
+    "StorageTechnologyConfig",
+    "ConversionTechnologyConfig"
 ]
