+++ zen_creator/elements/element.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, ClassVar, Type
 if TYPE_CHECKING:
     from zen_creator.model import Model
     from zen_creator.utils.attribute import Attribute
-    from zen_creator.utils.default_config import Config
+    from zen_creator.utils.config import Config
 import json
 from pathlib import Path
 
@@ -17,7 +17,7 @@ from zen_creator.utils.registry import Registry
 logger = logging.getLogger(__name__)
 
 
-class Element(Registry["Element"], ABC):
+class Element(ABC, Registry["Element"], is_base_registry = True):
     """Base class for all elements in the ZEN model.
 
     This class provides the foundation for carriers, technologies, and other
